+++ src/main/java/io/supertokens/pluginInterface/session/noSqlStorage/SessionInfoWithLastUpdated.java
@@ -23,9 +23,10 @@ public class SessionInfoWithLastUpdated extends SessionInfo {
     public String lastUpdatedSign;
 
     public SessionInfoWithLastUpdated(String sessionHandle, String userId, String refreshTokenHash2,
-            JsonObject userDataInDatabase, long expiry, JsonObject userDataInJWT, long timeCreated,
-            String lastUpdatedSign) {
-        super(sessionHandle, userId, refreshTokenHash2, userDataInDatabase, expiry, userDataInJWT, timeCreated);
+            JsonObject userDataInDatabase, long expiry, JsonObject userDataInJWT, JsonObject grantPayload,
+            long timeCreated, String lastUpdatedSign) {
+        super(sessionHandle, userId, refreshTokenHash2, userDataInDatabase, expiry, userDataInJWT, grantPayload,
+                timeCreated);
         this.lastUpdatedSign = lastUpdatedSign;
     }
 }